NFC: Remove build flag usage from Makefile
With the android upgrade requirement of separating
the build enablement rules from common projects and
adding them to tech team specific projects, new Makefile
has been added to the NFC system module.

The added Makefile defines the build rules which takes
care of the built PACKAGES and JARS to be included.

Removed build flag usage from the existing Makefile as
the built JAR will be included in PRODUCT_BOOT_JARS
on the basis of the new Makefile added as part of the
android upgarde requirement.

Change-Id: I0b8ba6368611ee7105a06efa1970479b9617a279